[design] goods 제품 이미지 클릭시 제품상세정보 볼 수 있게 추가
